+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - New function to load the user IDs who have retweeted a given tweet (`retweeters_of`)
 - New function to load the posts liked by a given user (`liked_by`)
 - New module `search` to contain all the tweet-search structs and methods
+- New functions to retweet and unretweet statuses (`retweet`, `unretweet`)
 
 ### Changed
 - Moved `UserID` into the user module